Update network.py
Confirmed that inception_v3 has GlobalAveragePooling2D. Made it so that the layers can be unfrozen by passing in an array with a list of layers to be unfrozen. If there is nothing in the array, all layers are unfrozen. Note that it is not possible in this implementation to freeze the last two layers (which we wouldn't want anyway, as they were just created).